Chore : 실종자 정보 등록 영역 컴포넌트화
